refactor(contracts): extract RustService trait + RustCallError
Step 1 of moving acp/managed_agent out of the kernel crate.

contracts is a leaf rlib (no dep on kernel). Moving the trait + its
error type there means an out-of-kernel service crate can implement
RustService without pulling kernel as a dep -- prerequisite for the
agent_runtime peer crate that owns acp/ + managed_agent/ in the
follow-up commit.

* rust/contracts/src/rust_service.rs   -- new (~80 LOC)
* rust/contracts/src/lib.rs            -- pub mod + re-export
* rust/kernel/src/service_registry.rs  -- delete the local copy,
                                          re-export from contracts so
                                          existing crate::service_registry::
                                          imports keep compiling

Verification:
  cargo clippy -p kernel --tests --all-features -- -D warnings  clean
  cargo test -p kernel --lib service_registry        9/9 passed
  cargo test -p kernel --lib acp::service::tests::dispatch  6/6 passed
